[pg15-upgrade] fix: set per-db catalog mode flag default to true
Summary:
DB catalog version mode isn't implemented for the upgrade process, but it's the default now in the master branch and the pg15 branch. To work around the lack of support during the upgrade, flag ysql_enable_db_catalog_version_mode's default value was set to false in pg15-upgrade merge commit d9cd490623633c061eeeeadb45238a82d8e563b5, but some non-upgrade tests are starting to depend on the new default flag value.

Instead of changing the flag default globally, restore the default to be true, same as on the other branches, but set the value to false on the command line for upgrade tests during the upgrade.

Test Plan:
Jenkins

  ./yb_build.sh release --sj --skip_tests
  ./yb_build.sh release reinitdb
  ./yb_build.sh release --java-test org.yb.pgsql.TestPgRegressAnalyze
  ./yb_build.sh release --java-test org.yb.pgsql.TestPgSequences
  pg15_tests/get_shell_test_specs.sh | grep upgrade | pg15_tests/run_tests.sh

Reviewers: jason

Reviewed By: jason

Subscribers: ybase, yql

Differential Revision: https://phorge.dev.yugabyte.com/D36935