[pg15] fix: avoid using incorrect MyProc in CVBroadcast cleanup function
Summary:
D36847 / 632949c9556df19e10bac683869b119a87c978fb broke the TestPgConnection tests.

`CleanupProcSignalStateInternal` is supposed to be able to be called from a process for itself, or from the postmaster for a recently killed process. However, it called the function `ConditionVariableBroadcast`, which relies on the static variable MyProc.

The fix is to explicitly pass in the PGPROC that we want to use for the `ConditionVariableBroadcast` function, by making `ConditionVariableBroadcast` a wrapper for a new function `YbConditionVariableBroadcastForProc`  - mostly the same except for a few changes to make it work for the given process.

Test Plan:
```
./yb_build.sh --java-test 'org.yb.pgsql.TestPgConnection'
```

Reviewers: aagrawal, kramanathan

Reviewed By: aagrawal

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D36977