[#23098] DocDB: yb-admin commands to dump and update sys catalog catalog entities
Summary:
`yb-master` stores all its data in the `sys_catalog` tablet as CatalogEntity objects. CatalogEntity
objects are stored as key value pairs where the key is `SysRowEntryType`,<entity_id> and the value
is a serialized PB message.

It's rare, but there have been cases where the CatalogEntity data gets corrupted due to bugs. When
this happens, yb workflows can get stuck or `yb-master` can enter a FATAL crash loop.
The sys-catalog-tool has the capability to dump `yb-master` data from the SST file, but it can only
be used when `yb-master` is not running. In most cases we have a healthy `yb-master`, but some stuck
workflows due to bad state of tables or xCluster and stopping the master is not an option
(and dangerous). The tool does not support updates, and even if it did, we would have to run it on
all master nodes, since it works on the SST files.

This change adds `emergency_repair_mode` flag to `yb-master`. When this is enabled, the master
(`sys_catalog`) leader will not load up `CatalogManager` which is useful when bad data is causing it
to crash loop. DumpSysCatalogEntries and WriteSysCatalogEntry are the only RPCs that can be run
against the leader in this mode. All other RPCs will be blocked by ScopedLeaderSharedLock which is
only initialized on a successful `CatalogManager` load. The writes performed via 
WriteSysCatalogEntry are RAFT replicated, so the change will get persisted across all masters. 
WriteSysCatalogEntry can only be used in `emergency_repair_mode` to prevent other master
bg/heartbeat activity from further corrupting the data.

New yb-admin command `dump_sys_catalog_entries` and `write_sys_catalog_entry` invoke the 
DumpSysCatalogEntries and WriteSysCatalogEntry RPC against the master leader. The insert and update 
options of `write_sys_catalog_entry` validate the schema of the new data. The update option dumps
the latest data from `sys_catalog` and shows a diff of the changes being made.

NOTE: `write_sys_catalog_entry` command is meant to be used in extreme situations when no other
mitigation is possible. It is very dangerous and can lead to corruption, data loss, and
unavailability. This has been marked as hidden, so that it does not show up in the yb-admin help.
In order to correctly map catalog entity types to their PBs introduce the `CATALOG_ENTITY_TYPE_MAP` 
macro. When a new `SysRowEntryType` only this macro needs to be updated, and it will handle all
other switch cases/ maps in the system automatically.

Notable considerations:

 - The serialization/deserialization of the proto to/from human-readable format is done by
 `yb-master` in order to guarantee its correctness. We don't do this in yb-admin since users can
 run a different version of it, or it can be built/compiled differently, leading to issues.
 - We use the proto debug serialization format instead of JSON since this is the only official
 supported format.

**Upgrade/Rollback safety:**
New proto messages are used only by yb-admin, which is always running the same version as
`yb-master`.

Fixes #23098
Jira: DB-12032

Test Plan:
AdminCliTest.TestDumpSysCatalogEntryInNonEmergencyMode
AdminCliTest.TestDumpSysCatalogEntryInEmergencyMode
AdminCliTest.BlockWriteToSysCatalogEntryInNonEmergencyMode
AdminCliTest.TestInsertDeleteSysCatalogEntry
AdminCliTest.TestUpdateSysCatalogEntry

Reviewers: slingam, xCluster, loginov, zdrudi

Reviewed By: loginov, zdrudi

Subscribers: ybase

Differential Revision: https://phorge.dev.yugabyte.com/D36262