[PLAT-14705] Fixing yb_platform_backup.sh for custom replicated storage path
Summary: Version checking was broken if replicated had set a custom storage directory because we were always looking inside of data dir but inside the container it is always /opt/yugabyte. For yba-installer only the data dir is exactly where we need to look for the version_metadata and it could change.

Test Plan:
Replicated with custom storage directory, take backup with version checking enabled.
Restore same backup to the custom storage YBA

Reviewers: dshubin, sanketh

Reviewed By: dshubin

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D36946