[#23428] docdb: Remove non-transactional snapshot code
Summary:
Transactional snapshots have been supported (and the default) since D8100 and D8349. The non-transactional snapshot flow is untested and clutters the transactional snapshots. To the best of our knowledge, it is unused. This diff removes the non-transactional snapshot code.

**Upgrade/Rollback safety:**
This diff deprecates the `transaction_aware` field in `CreateSnapshotRequestPB` and the `current_snapshot_id` field in `ListSnapshotsResponsePB`. Since masters are upgraded before tservers, the `transaction_aware` aware field will always be ignored (the new behavior) on an upgrade / downgrade. The `current_snapshot_id` field is only set for non-transactional snapshots (which are not being used), so that field's behavior does not change.
Jira: DB-12349

Test Plan: Existing tests

Reviewers: zdrudi, hsunder, loginov

Reviewed By: zdrudi, hsunder, loginov

Subscribers: loginov, ybase

Differential Revision: https://phorge.dev.yugabyte.com/D37010