[#21510] YSQL: Fix the unit test testSeekNextEstimationIndexScan which is failing during Connection Manager
Summary:
The UT `org.yb.pgsql.TestPgCostModelSeekNextEstimation#testSeekNextEstimationIndexScan` is failing because of a different set of results with and without connection manager.

| Metric name | With connection manager | Without connection manager
| **rocksdb_number_db_seek** | 4 | 2
| **rocksdb_number_db_next** | 10 | 4

The potential cause of the issue is due to cache misses resulting in higher number of seeks and nexts with connection manager. The issue has to be debugged further to have a RCA conclusively. We have also observed that only the first EXPLAIN command run has a disparity in the results, the subset runs has the results similar to the one without connection manager. The change here is to have a run of EXPLAIN command when connection manager is enabled followed by the tests that are running today. Also, a separate ticket would be filed for further debugging of the issue.

The ticket for further RCA is being tracked here - https://github.com/yugabyte/yugabyte-db/issues/23423
Jira: DB-10394

Test Plan:
Run the test `./yb_build.sh --enable-ysql-conn-mgr-test --java-test 'org.yb.pgsql.TestPgCostModelSeekNextEstimation'` and verify that the test runs successfully with connection manager

Run the test `./yb_build.sh --java-test 'org.yb.pgsql.TestPgCostModelSeekNextEstimation'` and verify that the test runs successfully with connection manager disabled.

Jenkins: test regex: .*TestPgCostModelSeekNextEstimation.*, enable connection manager

Reviewers: asrinivasan, mkumar, rbarigidad, skumar, nkumar

Reviewed By: asrinivasan

Differential Revision: https://phorge.dev.yugabyte.com/D37035