[#23401] YSQL: Graceful error if syscache is accessed in DocDB
Summary:
Catalog cache is not available in DocDB, so the code that makes catalog
cache lookups should not be executed there. In other words,
those functions and expressions should not be pushable.

We can't test 100% of function for pushability, and can't future proof
it, so it is quite possible if some rarely used function we allowed to
be pushed down makes such lookups. Severity of that kind of error is
quite high, it causes tserver crash by failed assertion in debug build
or by segmentation fault in release builds. Such an error may disrupt
production environment and hard to RCA, since it is not clear, which
query running in the cluster has caused the crash.

Here we are adding a check for multithreaded mode in the catalog lookup
functions and gracefully error out if it is the case. That way only the
offending query will fail, and it will be immediately clear how to
mitigate the problem (disable expression pushdown for the query).
Jira: DB-12319

Test Plan:
There are no known cases of pushing down code making catalog lookups,
and we should promply fix, as soon as we aware of them.
So, to test, we
- make sure existing tests are not broken;
- revert last known fix of that kind of issue and check if error is now graceful:
```
git revert 403071d
... build and run ...
yugabyte=# select * from test;
  data
---------
 foo bar
(1 row)

yugabyte=# select * from test where 'foo' = ANY(string_to_array(data, ' '));
ERROR:  Catalog cache lookup is not allowed in multithread mode
```

Reviewers: jason, tnayak

Reviewed By: jason

Subscribers: yql

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D37064