[BACKPORT 2024.1][#23260] CDCSDK: Propagate stream expiry & GC error to walsender
Summary:
**Backport description:**
No merge conflicts

**Original description:**
Original commit: 5ed864d821e6284599939747aabb42a32fc1b852 / D36756
GetConsistentChanges RPC, at the moment, does not propagate any error to the walsender. But, with this diff, we will now propagate the following errors to the walsender as we cannot recover from these errors and the only solution is to create a new CDC stream.

  - Stream has expired
  - Intents have been GC'ed

The consequence of sending these errors to walsender is it will lead to a crash in walsender, thereby, propagating the error further to the actual CDC client (eg: YugabyteDB connector). Such errors are not to be retried by CDC clients as these are non-recoverable errors as mentioned above.
Jira: DB-12188

Test Plan:
Jenkins: urgent, test regex: .*CDC.*

./yb_build.sh --cxx-test cdcsdk_consumption_consistent_changes-test --gtest_filter CDCSDKConsumptionConsistentChangesTest.TestIntentGC

./yb_build.sh --cxx-test cdcsdk_consumption_consistent_changes-test --gtest_filter CDCSDKConsumptionConsistentChangesTest.TestStreamExpiry

Reviewers: asrinivasan, skumar, stiwary

Reviewed By: asrinivasan

Subscribers: ycdcxcluster

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D37092