[PLAT-13601] Add current platform host URL to YBA footer
Summary:
This diff adds the local YBA platform URL to the footer if
the current browser origin is different than the current local
platform address (ex. in the case of load balancer setup).

This is useful because users may want to know the current instance's
address to ssh into it.

Test Plan:
Verify that the YBA platform URL is only shown when the current window
origin differs from the local instance address.
{F273047}

Reviewers: muthu, rmadhavan, sanketh

Reviewed By: rmadhavan

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D37006