Fix Issue 1988: How to update a property which is a keyword (#2005) (#2015)
Fixed issue 1988: Reserved Keyword Handling: How to update a property
which is a keyword. Basically, how to use a keyword, aka reserved
word, in a query.

Added the appropriate scanner and parser rules to allow keywords to
be back ticked and used in queries as their name. For example, like
in the following commands -

    MATCH (n:Person { id: 123 }) SET n.`match` = 'matched' RETURN n;
    MATCH (n:Person { id: 123 }) SET n.`create` = 'created' RETURN n;

The changes did not break any regression tests.
Added additional regression tests.