[#21202] YSQL: Merge Connection Manager logs with Java UT logs
Summary:
This diff helps to include connection manager debug and query logs within detective reports. As of today, the test suite is run with a flag `logtostderr` which allows master and tserver logs to be pushed onto stderr. stderr is then piped to stdout, which is then used to then populate the file where Java UT logs are stored for each UT that is run. Since connection manager is spawned as a child process of tserver processes, we can use the same thread that pipes stdout to the java UT log file to have connection manager logs populated as well, all we would need to do is push connection manager logs to either stdout or stderr.

This patch re-uses the `logtostderr` flag to push connection manager logs to stdout (already in-built feature within odyssey) as a quick-fix to help populate detective reports with connection manager logs

Jira: DB-10132

Test Plan:
Jenkins: enable connection manager, all tests

Ensure connection manager logs are visible in detective runs

Reviewers: asrinivasan, skumar

Reviewed By: asrinivasan

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D37074