[pg15] fix: null pointer dereference in YbGetDdlMode
Summary:
Upstream PG commit f4a3fdfbdcd3763c42111318d004c2e90d072021 modifies parse_utilcmd.c:transformIndexConstraint to generate an extra AT_SetNotNull command when we add a primary key index to a table. This alterTableCmd has a null `def` pointer, which we dereference in YbGetDdlMode, resulting in a segfault.

This revision checks that the pointer is not null before dereferencing it.

Test Plan:
```
./yb_build.sh --java-test 'org.yb.pgsql.TestYbGetDdlMode`
```

Reviewers: jason, tfoucher

Reviewed By: jason

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D37007