Merged PR 1423558: [pgmongo] Defensive check for epoch conversion to pg_timestamp
### Does this PR have any customer impact?

### Type (Feature, Refactoring, Bugfix, DevOps, Testing, Perf, etc)
bugfix

### Does it involve schema level changes? (Table, Column, Index, UDF, etc level changes)

### Are you introducing any new config? If yes, do you have tests with and without them being set?

### Release Note (Refer [Template](../docs/release_notes/pgmongo_template.md))

### Description

adding a defensive check before
Generally all the drivers or clients have the check to only allow valid ranges of datetime/timestamp as input.
But , on a specific case if cx tries to convert to date a number or long using $todate function and then tries to convert it to string from date object then we were giving internal error.

As of now , have added a defensive check to fail fast on error before calling pg function

----
#### AI description  (iteration 1)
#### PR Classification
Bug fix to add a defensive check for epoch conversion to PostgreSQL timestamp.

#### PR Summary
This pull request introduces a defensive check to prevent errors during the conversion of Unix epoch milliseconds to PostgreSQL timestamps.
- `oss/pg_helio_core/include/utils/date_utils.h`: Added a range check to validate the seconds before conversion to `timestamptz`.
- `pgmongo/src/test/regress/sql/bson_aggregation_date_operators_tests.sql`: Added test cases to verify the defensive check.
- `pgmongo/src/test/regress/expected/bson_aggregation_date_operators_tests.out`: Updated expected output to reflect the new error messages for invalid conversions.