[#22373] docdb: Backward scans: optimize performance for multiple column updates
Summary:
With fast backward scan enabled (gflag `use_fast_backward_scan`) and in case of multiple updates
for the same column (which are not yet compacted down to a single key-value in RocksDB), current
implementation will iterate through all the updates from the oldest one to the newest one using a
`Prev` call. But in such cases it is not effective to loop through all the updates and the better
choice is to use a `Seek` call to the particular key's column if the suitable record is not reached
via reasonable number of `Prev` calls.

The diff implements that logic: if the iterator is not moved from the current column by using
X `Prev` calls, then a `Seek` is triggered to reach the very newest (from read time perspective)
update of the current column. The number of `Prev` calls needed to jump out a column is set to 2.
Gflag `max_prevs_to_avoid_seek` is used to control/tweak that value.
Jira: DB-11274

Test Plan:
./yb_build.sh --cxx-test=pg_single_tserver-test --gtest_filter=*/PgFastBackwardScanOnlyTest.ColumnMultipleUpdates/kNone
./yb_build.sh --cxx-test=pg_single_tserver-test --gtest_filter=*/PgFastBackwardScanOnlyTest.ColumnMultipleUpdates/kV1
./yb_build.sh --cxx-test=pg_single_tserver-test --gtest_filter=*/PgFastBackwardScanOnlyTest.ColumnMultipleUpdates/kV2

./yb_build.sh --cxx-test=pg_single_tserver-test --gtest_filter=*/PgFastBackwardScanOnlyTest.ColumnMultipleUpdatesColocated/kNone
./yb_build.sh --cxx-test=pg_single_tserver-test --gtest_filter=*/PgFastBackwardScanOnlyTest.ColumnMultipleUpdatesColocated/kV1
./yb_build.sh --cxx-test=pg_single_tserver-test --gtest_filter=*/PgFastBackwardScanOnlyTest.ColumnMultipleUpdatesColocated/kV2

Reviewers: sergei, rthallam

Reviewed By: sergei, rthallam

Subscribers: ybase

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D37203