[#22925] docdb: Refactor TSDescriptor to a sys catalog entity.
Summary:
This diff refactors the in-memory TSDescriptors to use the new SysServerEntryPB, following the
conventions and patterns of the other sys catalog entities. This diff does _NOT_ add sys catalog
writes, the TSDescriptor data is still purely transient in memory. There should be no functional
changes with this diff. The motivation is to enable further diffs that add flag-guarded logic to
persist TSDescriptors to the sys catalog.
Jira: DB-11843

Test Plan: Existing tests.

Reviewers: asrivastava, xCluster, hsunder

Reviewed By: asrivastava

Subscribers: ybase, esheng, yql, slingam

Differential Revision: https://phorge.dev.yugabyte.com/D36510