[#22876][#22835][#22773] CDCSDK: Add null checks & remove optimisation in table removal from CDC stream
Summary:
This diff make the following changes:

 # The table (non-eligible/user-created) removal codepath as well as addition of children tablets on split fetches tableInfoPtr/CDCStreamInfoPtr from master. At some of the places, we were either missing null checks which have been added
 # Removes the optimization in `UpdateCheckpointForTabletEntriesInCDCState()` that fetches TabletInfo only for a filtered set of tablet_ids when a particular table is being removed. This set used to only contain CDC state entries for the table being removed. Due to this optimisation, there was a race condition with drop table where this table removal codepath removes the table from stream metadata and the drop table cleanup codepath deletes the relevant state table entries. The race is described below:

Current algorithm for cleanup of user-created table (yb-admin command)
1. Check if stream is non-null & not in DELETING state.
2. Check if table is non-null & not in DELETING state
3. Update cdc state table entries

   i) Get tablets belonging to the table to be removed from master

    ii) Filter cdc state table entries whose tablets are part of the above collected tablet list

    iii) Get TabletInfo of filtered state table entries and check the table_id of the tablet

    iv) Iterate over the TabletInfo(s) and update checkpoint to max if table_id in TabletInfo doesnt belong to stream metadata.

    v) Delete colocated table state table entry

4. Remove user table from stream metadata

In Step 3, i) was the optimisation. The race here is suppose between step-2 & step-3, the table being removed was dropped. When we execute i) of step-3, we would receive 0 tablets from master for the table being removed. Because of 0 tablets, we wont be able to filter out any state table entries. Hence, essentially, we will not be updating any state table entry. But in step-4, we will remove the table from stream metadata. Now when the CDC drop table cleanup bg thread executes, it will find these state table entries that doesnt belong to any tables in stream metadata and hence will directly delete them. Although the end result is as expected, but we would like to avoid these partial cleanup from codepaths.

The updated algorithm for updating the checkpoint of a state table entry is as follows:

We will now scan the entire state table and fetch the TabletInfo for all the tablets in cdc state table. A state table entry can qualify for checkpoint update based on if a particular table is being removed or if all the state table entries are just being validated and synced with tables in stream metadata.

Case-1 : Removal of a particular table. Tablet has to satisfy either of the two conditions for checkpoint update:

  - If the table being removed is not a colocated table, therefore the tablet exclusively belongs to the table being removed.
  - If the table being removed is a colocated table, then all the other colocated tables on the tablet should not be present in stream metadata.

Case-2: Validation and sync of cdc state entries

  - A state table entry can only qualify for checkpoint update if it belongs to none of the tables present in stream metadata.
Jira: DB-11778, DB-11733, DB-11676

Test Plan:
Jenkins: urgent, test regex: .*TestNonEligibleTableCleanupWithDropTable.*
./yb_build.sh --cxx-test cdcsdk_ysql-test --gtest_filter CDCSDKYsqlTest.TestUserTableCleanupWithDropTable
./yb_build.sh --cxx-test cdcsdk_ysql-test --gtest_filter CDCSDKYsqlTest.TestUserTableCleanupWithDeleteCDCStream
./yb_build.sh --cxx-test cdcsdk_ysql-test --gtest_filter CDCSDKYsqlTest.TestNonEligibleTableCleanupWithDropTable
./yb_build.sh --cxx-test cdcsdk_ysql-test --gtest_filter CDCSDKYsqlTest.TestNonEligibleTableCleanupWithDeleteStream
./yb_build.sh --cxx-test cdcsdk_ysql-test --gtest_filter CDCSDKYsqlTest.TestColocatedUserTableRemovalFromCDCStream

Reviewers: xCluster, hsunder, skumar, stiwary, sumukh.phalgaonkar, asrinivasan

Reviewed By: skumar

Subscribers: ybase

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D37135